pr-checks: extract output utilities
