+++ pkg/cmd/pr/checks/checks.go
@@ -2,8 +2,6 @@ package checks
 
 import (
 	"fmt"
-	"sort"
-	"time"
 
 	"github.com/MakeNowJust/heredoc"
 	"github.com/cli/cli/v2/internal/ghrepo"
@@ -91,109 +89,18 @@ func checksRunWebMode(opts *ChecksOptions) error {
 	return opts.Browser.Browse(openURL)
 }
 
-func addRow(tp utils.TablePrinter, isTerminal bool, cs *iostreams.ColorScheme, o check) {
-	elapsed := ""
-	zeroTime := time.Time{}
-
-	if o.StartedAt != zeroTime && o.CompletedAt != zeroTime {
-		e := o.CompletedAt.Sub(o.StartedAt)
-		if e > 0 {
-			elapsed = e.String()
-		}
-	}
-
-	mark := "✓"
-	markColor := cs.Green
-	switch o.Bucket {
-	case "fail":
-		mark = "X"
-		markColor = cs.Red
-	case "pending":
-		mark = "*"
-		markColor = cs.Yellow
-	case "skipping":
-		mark = "-"
-		markColor = cs.Gray
-	}
-
-	if isTerminal {
-		tp.AddField(mark, nil, markColor)
-		tp.AddField(o.Name, nil, nil)
-		tp.AddField(elapsed, nil, nil)
-		tp.AddField(o.Link, nil, nil)
-	} else {
-		tp.AddField(o.Name, nil, nil)
-		tp.AddField(o.Bucket, nil, nil)
-		if elapsed == "" {
-			tp.AddField("0", nil, nil)
-		} else {
-			tp.AddField(elapsed, nil, nil)
-		}
-		tp.AddField(o.Link, nil, nil)
-	}
-
-	tp.EndRow()
-}
-
 func checksRun(opts *ChecksOptions) error {
 	checks, meta, err := collect(opts)
 	if err != nil {
 		return fmt.Errorf("cannot collect checks information: %w", err)
 	}
 
-	fmt.Printf("%+v\n", checks)
-
 	isTerminal := opts.IO.IsStdoutTTY()
 	cs := opts.IO.ColorScheme()
 
-	summary := ""
-	if meta.Failed+meta.Passed+meta.Skipping+meta.Pending > 0 {
-		if meta.Failed > 0 {
-			summary = "Some checks were not successful"
-		} else if meta.Pending > 0 {
-			summary = "Some checks are still pending"
-		} else {
-			summary = "All checks were successful"
-		}
-
-		tallies := fmt.Sprintf("%d failing, %d successful, %d skipped, and %d pending checks",
-			meta.Failed, meta.Passed, meta.Skipping, meta.Pending)
-
-		summary = fmt.Sprintf("%s\n%s", cs.Bold(summary), tallies)
-	}
-
-	if isTerminal {
-		fmt.Fprintln(opts.IO.Out, summary)
-		fmt.Fprintln(opts.IO.Out)
-	}
-
-	tp := utils.NewTablePrinter(opts.IO)
-
-	sort.Slice(checks, func(i, j int) bool {
-		b0 := checks[i].Bucket
-		n0 := checks[i].Name
-		l0 := checks[i].Link
-		b1 := checks[j].Bucket
-		n1 := checks[j].Name
-		l1 := checks[j].Link
-
-		if b0 == b1 {
-			if n0 == n1 {
-				return l0 < l1
-			}
-			return n0 < n1
-		}
-
-		return (b0 == "fail") || (b0 == "pending" && b1 == "success")
-	})
-
-	for _, o := range checks {
-		addRow(tp, isTerminal, cs, o)
-	}
-
-	err = tp.Render()
+	err = printTable(isTerminal, cs, meta, checks, opts)
 	if err != nil {
-		return err
+		return fmt.Errorf("cannot print table: %w", err)
 	}
 
 	if meta.Failed+meta.Pending > 0 {
